requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.3.0
black==19.10b0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gspread==3.1.0
gunicorn==19.9.0
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
multidict==4.5.2
oauth2client==4.1.3
oauthlib==3.1.0
pathspec==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pyee==6.0.0
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
slackclient==2.3.0
slacker==0.13.0
slackeventsapi==2.1.0
toml==0.10.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
websocket-client==0.56.0
Werkzeug==0.16.0
yarl==1.3.0
